+++ src/Request.php
@@ -62,6 +62,16 @@ class Request extends Message
      */
     private $_contentMap = [];
 
+    /** 
+     * @var float stores the starttime of the current request with microsecond-precession
+     */
+    private $_startTime;
+
+    /**
+     * @var float stores the seconds of how long does it take to get a response
+     */
+    private $_timeElapsed;
+
 
     /**
      * Sets target URL.
@@ -448,6 +458,7 @@ class Request extends Message
         $event = new RequestEvent();
         $event->request = $this;
         $this->trigger(self::EVENT_BEFORE_SEND, $event);
+        $this->_startTime = microtime(true);
     }
 
     /**
@@ -458,12 +469,24 @@ class Request extends Message
      */
     public function afterSend($response)
     {
+        $this->_timeElapsed = microtime(true)-$this->_startTime;
         $this->client->afterSend($this, $response);
 
         $event = new RequestEvent();
         $event->request = $this;
         $event->response = $response;
         $this->trigger(self::EVENT_AFTER_SEND, $event);
+    }
+
+    /**
+     * Return the response time in seconds
+     *
+     * @return float the seconds elapsed from request to response
+     * @since 2.0.12
+     */
+    public function responseTime()
+    {
+        return $this->_timeElapsed;
     }
 
     /**